README.md: include example for python3.packageOverrides
The python3 `packageOverrides` argument is a real footgun; it's supposed to be
an overlay, but instead of `extendOverlay`ing it tons of things in nixpkgs just
clobber the preexisting value.  This example shows off how infusions won't let
you do that unintentionally.  They also make it way easier to access this
functionality.