change type of (nonpublic) unique-marker from [<function>] to [[<throw>]]
infuse.nix has an internal, nonpublic "missing value" marker, which uses
sternensseeman@'s pointer equality trick to implement gensym().

This commit changes the missing value marker from

  [(_: throw "message")]

to

  [[(throw "message")]]

In other words, instead of a list containing a function which throws, the marker
is now a list containing a list containing a (thunked) throw.

This change should not be externally visible.

Signed-off-by: Adam Joseph <adam@westernsemico.com>