change strategy for dealing with missing arguments
The file `doc/design-notes-on-missing-attributes.md` lists two
possible strategies for dealing with missing attributes: `__functor`
or @sternenseeman-style incomparables (gensym).

Prior to this commit, the `__functor` strategy was used.  It turns out
that it is impossible to satisfy the left identity law using that
strategy, so this commit switches to the incomparables/gensym strategy.

See the test case added by this commit for a concrete example of why
the `__functor` strategy cannot work.

This is an externally-visible breaking change for downstream users who
were defining their own sugars.  Custom sugars are a very advanced
feature; at this point there are very few such users.

Signed-off-by: Adam Joseph <adam@westernsemico.com>