README.md: use `flip pipe` instead of undefined `compose`
Prior to this commit, the semantics section of `README.md` referred to
`compose`, which isn't in nixpkgs/lib.  And it wouldn't have worked with the
Haskell version of `compose`.

This commit fixes the mistake by rewriting the examples to use `flip pipe`.

Thanks to @lucus and @sterni for noticing this.