when printing error messages, wrap toPretty in builtins.tryEval
This commit does not produce any eval-detectable behavior changes.  The only
changes it causes are in the error messages and trace printouts when an error
occurs.

The infuse library tries to be helpful by printing as much information as
possible about the location where an error occurred.  This is extremely helpful
in lazy languages like nix, where there are few good debugging tools.

The extra information printed includes the attrpath at which an error occurred,
and for type errors the value which had an unexpected type.  If the value with
an unexpected type is in fact a `throw` value, the error message itself will
trigger a further error, which will be the last printout from `--show-trace`.
This is generally not as helpful as simply reporting the value-with-wrong-type
as "<<throw>>".

This commit wraps a `builtins.tryEval` invocation around `toPretty` to give the
more-helpful behavior.

Signed-off-by: Adam M. Joseph <adam@westernsemico.com>